+++ OrganizationSearch/FindOrganization.test.js
@@ -1,6 +1,4 @@
-import React from 'react';
-import { render, act } from '@testing-library/react';
-
+import { render, act } from '@folio/jest-config-stripes/testing-library/react';
 import {
   FindRecords,
   PLUGIN_RESULT_COUNT_INCREMENT,